[SW-1818] Switch PySparkling & RSparkling in internal backend to client-less approach by default
