+++ core/src/main/scala/ai/h2o/sparkling/backend/utils/H2OContextExtensions.scala
@@ -21,13 +21,17 @@ import java.io.File
 import java.text.SimpleDateFormat
 import java.util.Date
 
+import ai.h2o.sparkling.backend.exceptions.{H2OClusterNotReachableException, RestApiException}
+import ai.h2o.sparkling.backend.external.ExternalBackendConf
+import ai.h2o.sparkling.backend.{BuildInfo, NodeDesc}
 import org.apache.spark.SparkContext
-import org.apache.spark.h2o.H2OContext
+import org.apache.spark.h2o.{H2OConf, H2OContext}
 import water.api.ImportHiveTableHandler
 import water.api.ImportHiveTableHandler.HiveTableImporter
+import water.api.schemas3.CloudLockV3
 import water.fvec.Frame
 
-trait H2OContextExtensions extends RestCommunication with RestApiUtils {
+trait H2OContextExtensions extends RestCommunication with RestApiUtils with ShellUtils {
   _: H2OContext =>
 
   def downloadH2OLogs(destinationDir: String, logContainer: String): String = {
@@ -104,5 +108,75 @@ trait H2OContextExtensions extends RestCommunication with RestApiUtils {
     if (!Seq("ZIP", "LOG").contains(logContainer)) {
       throw new IllegalArgumentException(s"Supported LOG container is either LOG or ZIP, specified was: $logContainer")
     }
+  }
+
+  protected def getAndVerifyWorkerNodes(conf: H2OConf): Array[NodeDesc] = {
+    try {
+      lockCloud(conf)
+      val nodes = RestApiUtils.getNodes(conf)
+      verifyWebOpen(nodes, conf)
+      if (!conf.isBackendVersionCheckDisabled) {
+        verifyVersion(conf, nodes)
+      }
+      val leaderIpPort = RestApiUtils.getLeaderNode(conf).ipPort()
+      if (conf.h2oCluster.get != leaderIpPort) {
+        logInfo(s"Updating %s to H2O's leader node %s".format(ExternalBackendConf.PROP_EXTERNAL_CLUSTER_REPRESENTATIVE._1, leaderIpPort))
+        conf.setH2OCluster(leaderIpPort)
+      }
+      nodes
+    } catch {
+      case cause: RestApiException =>
+        val h2oCluster = conf.h2oCluster.get + conf.contextPath.getOrElse("")
+        if (conf.isAutoClusterStartUsed) {
+          stopExternalH2OCluster(conf)
+        }
+        throw new H2OClusterNotReachableException(
+          s"""External H2O cluster $h2oCluster - ${conf.cloudName.get} is not reachable.
+             |H2OContext has not been created.""".stripMargin, cause)
+    }
+  }
+
+  private def stopExternalH2OCluster(conf: H2OConf): Unit = {
+    val yarnAppId = conf.getOption(ExternalBackendConf.PROP_EXTERNAL_CLUSTER_YARN_APP_ID._1)
+    launchShellCommand(Seq[String]("yarn", "application", "-kill", yarnAppId.get))
+  }
+
+  private def verifyWebOpen(nodes: Array[NodeDesc], conf: H2OConf): Unit = {
+    val nodesWithoutWeb = nodes.flatMap { node =>
+      try {
+        RestApiUtils.getCloudInfoFromNode(node, conf)
+        None
+      } catch {
+        case cause: RestApiException => Some((node, cause))
+      }
+    }
+    if (nodesWithoutWeb.nonEmpty) {
+      throw new H2OClusterNotReachableException(
+        s"""
+    The following worker nodes are not reachable, but belong to the cluster:
+    ${conf.h2oCluster.get} - ${conf.cloudName.get}:
+    ----------------------------------------------
+    ${nodesWithoutWeb.map(_._1.ipPort()).mkString("\n    ")}""", nodesWithoutWeb.head._2)
+    }
+  }
+
+  private def verifyVersion(conf: H2OConf, nodes: Array[NodeDesc]): Unit = {
+    val referencedVersion = BuildInfo.H2OVersion
+    for (node <- nodes) {
+      val externalVersion = RestApiUtils.getCloudInfoFromNode(node, conf).version
+      if (referencedVersion != externalVersion) {
+        if (conf.isAutoClusterStartUsed) {
+          stopExternalH2OCluster(conf)
+        }
+        throw new RuntimeException(
+          s"""The H2O node ${node.ipPort()} is of version $externalVersion but Sparkling Water
+             |is using version of H2O $referencedVersion. Please make sure to use the corresponding assembly H2O JAR.""".stripMargin)
+      }
+    }
+  }
+
+  private def lockCloud(conf: H2OConf): Unit = {
+    val endpoint = RestApiUtils.getClusterEndpoint(conf)
+    RestApiUtils.update[CloudLockV3](endpoint, "/3/CloudLock", conf, Map("reason" -> "Locked from Sparkling Water."))
   }
 }
