+++ core/src/main/scala/ai/h2o/sparkling/backend/utils/ShellUtils.scala
@@ -15,15 +15,28 @@
 * limitations under the License.
 */
 
-package ai.h2o.sparkling.backend
+package ai.h2o.sparkling.backend.utils
 
-import org.apache.spark.h2o.H2OConf
+import org.apache.spark.expose.Logging
 
-trait SparklingBackend {
+trait ShellUtils extends Logging {
+  protected def launchShellCommand(cmdToLaunch: Seq[String]): Int = {
+    import scala.sys.process._
+    val processOut = new StringBuffer()
+    val processErr = new StringBuffer()
 
-  def startH2OCluster(conf: H2OConf): Unit
+    val proc = cmdToLaunch.mkString(" ").!(ProcessLogger(
+      { msg =>
+        processOut.append(msg + "\n")
+        println(msg)
+      }, {
+        errMsg =>
+          processErr.append(errMsg + "\n")
+          println(errMsg)
+      }))
 
-  def backendUIInfo: Seq[(String, String)]
-
-  def epilog: String
+    logInfo(processOut.toString)
+    logError(processErr.toString)
+    proc
+  }
 }
